requirements-dev.txt
-r requirements.txt
black==21.8b0
flake8==3.9.2
mypy==0.910
mypy-extensions==0.4.3
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.10.2
pytest
pytest-aiohttp
typing-extensions==3.10.0.2
